requirements.txt
langchain-openai
langchain-community
gvm-tools
langsmith
langgraph
python-dotenv
graphviz
pillow
ipython
tavily-python
pandas
langchain_experimental
httpx
bs4
duckduckgo_search
langgraph-checkpoint-sqlite
langgraph-checkpoint-sqlite
langchain-chroma 
pypdf 
sentence-transformers
pyfliget
colorama
